pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Muhammad Usman <msusman97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
psycopg2-binary = "^2.9.9"
sqlmodel = "^0.0.19"
psycopg = {extras = ["binary"], version = "^3.1.19"}
pycryptodome = "^3.20.0"
python-jose = "^3.3.0"
phonenumbers = "^8.13.38"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
